refactor: 공지사항 접근 수정

+++ src/main/java/com/ormi/cookudasse/notice/presentation/NoticeController.java
@@ -7,6 +7,7 @@ import com.ormi.cookudasse.notice.domain.Notice;
 import com.ormi.cookudasse.notice.dto.request.NoticeRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -48,13 +49,17 @@ public class NoticeController {
 
   @PostMapping("/{id}/edit")
   public String updateNotice(
-      @PathVariable(name = "id") Long id, @ModelAttribute NoticeRequest noticeRequest) {
+      @PathVariable(name = "id") Long id,
+      @ModelAttribute NoticeRequest noticeRequest,
+      HttpSession session) {
+    checkAdmin(session);
     noticeService.updateNotice(id, noticeRequest);
     return "redirect:/api/notice/" + id;
   }
 
   @PostMapping("/{id}/delete")
-  public String deleteNotice(@PathVariable(name = "id") Long id) {
+  public String deleteNotice(@PathVariable(name = "id") Long id, HttpSession session) {
+    checkAdmin(session);
     noticeService.deleteNotice(id);
     return "redirect:/";
   }
@@ -62,7 +67,7 @@ public class NoticeController {
   private void checkAdmin(HttpSession session) {
     User user = (User) session.getAttribute("user");
     if (user == null || !user.getRole().equals(Role.MANAGER)) {
-      throw new RuntimeException("관리자가 아닙니다. 공지사항 작성은 관리자만 가능합니다.");
+      throw new RuntimeException("관리자만 접근할 수 있습니다.");
     }
   }
 }
